Fix critical bugs and update documentation
🐛 Bug Fixes:
- Fix TTL casting issue in config/oneclicklogin.php (add int cast)
- Fix ValidateConfigCommand to use correct config key 'ttl_minutes'
- Optimize MagicLinkManager::revoke() method for unused links only
- Add proper error message for revoke method

📚 Documentation Updates:
- Update all config file references from 'sharelink.php' to 'oneclicklogin.php'
- Fix config key references from 'default_expiration_minutes' to 'ttl_minutes'
- Correct service provider class name in all documentation files
- Ensure consistency between docs and actual implementation

✅ All tests passing (73/73 with 265 assertions)
✅ PHPStan level 9 compliance
✅ Laravel Pint style compliance

+++ config/oneclicklogin.php
@@ -4,10 +4,10 @@ declare(strict_types=1);
 
 return [
     // Time-to-live for magic links in minutes
-    'ttl_minutes' => env('ONECLICKLOGIN_TTL_MINUTES', 15),
+    'ttl_minutes' => (int) env('ONECLICKLOGIN_TTL_MINUTES', 15),
 
     // Maximum number of uses per magic link
-    'max_uses' => env('ONECLICKLOGIN_MAX_USES', 1),
+    'max_uses' => (int) env('ONECLICKLOGIN_MAX_USES', 1),
 
     // Route name for consuming magic links
     'signed_route_name' => 'oneclicklogin.consume',
